+++ routes/stream.route.js
@@ -1,6 +1,6 @@
 const routes = require('express').Router()
 const controller = require('../controllers/stream.controller')
 
-routes.get('/', controller.get)
+routes.post(controller.stream)
 
 module.exports = routes
